Split packet encoder into focused modules
