+++ rust/otap-dataflow/crates/contrib-nodes/Cargo.toml
@@ -31,6 +31,7 @@ prost.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
+crc32fast = { workspace = true, optional = true }
 linkme.workspace = true
 tokio.workspace = true
 tracing.workspace = true
@@ -130,7 +131,7 @@ contrib-exporters = [
     "kafka-exporter",
     "clickhouse-exporter",
 ]
-geneva-metrics-exporter = []
+geneva-metrics-exporter = ["dep:crc32fast"]
 clickhouse-exporter = [
     "dep:clickhouse",
     "dep:clickhouse-ext-arrow",
