+++ rust/otap-dataflow/crates/contrib-nodes/src/exporters/geneva_metrics_exporter/encoder/fixtures/MetricTimeStampProtocol6.bin
