+++ rust/otap-dataflow/crates/contrib-nodes/src/exporters/geneva_metrics_exporter/encoder/fixtures/MinMaxHistoExponentialMetricDouble.bin
