+++ rust/otap-dataflow/crates/contrib-nodes/src/exporters/geneva_metrics_exporter/encoder/mod.rs
@@ -1,0 +1,17 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//! Geneva Metrics ingestion protocol version 6 packet model and encoder.
+
+mod exemplar;
+mod histogram;
+mod model;
+mod numeric;
+mod packet;
+mod writer;
+
+pub use model::*;
+pub use packet::encode;
+
+#[cfg(test)]
+mod test_support;
