+++ rust/otap-dataflow/crates/contrib-nodes/src/exporters/geneva_metrics_exporter/mod.rs
@@ -1,6 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-//! ME-to-FE protocol version 6 model and encoding.
+//! Geneva Metrics ingestion protocol version 6 model and encoding.
 
-pub mod codec;
+pub mod encoder;
